SchemaCache struct (ala SchemaRegistry) (#205)
* To remove some duplication and make things a bit cleaner, we add a
`SchemaCache` struct, similar to the `SchemaRegistry` one, with the
properties that were hardcoded on the ZTF and DECAM alert workers.

* Then, we move the `alert_from_avro_bytes` method out of the
AlertWorker trait, and remove their survey-specific implementations from
all 3 surveys, to move them to the SchemaRegistry and newly added
SchemaCache struct. Now, surveys implement one of these 2 data
structures to handle schema's and reading data from avro.
