+++ src/alert/mod.rs
@@ -6,6 +6,6 @@ pub use base::{
     deserialize_mjd, deserialize_mjd_option, get_schema_and_startidx, run_alert_worker, AlertError,
     AlertWorker, AlertWorkerError, ProcessAlertStatus, SchemaRegistry, SchemaRegistryError,
 };
-pub use decam::{DecamAlertWorker, DECAM_DEC_RANGE};
-pub use lsst::{LsstAlertWorker, LSST_DEC_RANGE, LSST_SCHEMA_REGISTRY_URL};
-pub use ztf::{ZtfAlertWorker, ZTF_DECAM_XMATCH_RADIUS, ZTF_LSST_XMATCH_RADIUS};
+pub use decam::{DecamAlert, DecamAlertWorker, DECAM_DEC_RANGE};
+pub use lsst::{LsstAlert, LsstAlertWorker, LSST_DEC_RANGE, LSST_SCHEMA_REGISTRY_URL};
+pub use ztf::{ZtfAlert, ZtfAlertWorker, ZTF_DECAM_XMATCH_RADIUS, ZTF_LSST_XMATCH_RADIUS};
